Rename `type_utils` to `types`
