+++ sentinelhub/api/catalog.py
@@ -9,7 +9,7 @@ from ..base import FeatureIterator
 from ..data_collections import DataCollection, OrbitDirection
 from ..geometry import CRS, BBox, Geometry
 from ..time_utils import parse_time, parse_time_interval, serialize_time
-from ..type_utils import JsonDict, RawTimeIntervalType, RawTimeType
+from ..types import JsonDict, RawTimeIntervalType, RawTimeType
 from .base import SentinelHubService
 from .utils import remove_undefined
